Create showList module and export it to index.js
